[tests] extract pipeline → mongo tests in a separate file
